[collector]feature:Add monitoring parameters for Redis Cluster and Sentinel
