+++ collector/src/main/resources/META-INF/services/com.usthe.collector.collect.AbstractCollect
@@ -3,7 +3,7 @@ com.usthe.collector.collect.http.SslCertificateCollectImpl
 com.usthe.collector.collect.database.JdbcCommonCollect
 com.usthe.collector.collect.icmp.IcmpCollectImpl
 com.usthe.collector.collect.jmx.JmxCollectImpl
-com.usthe.collector.collect.redis.RedisSingleCollectImpl
+com.usthe.collector.collect.redis.RedisCommonCollectImpl
 com.usthe.collector.collect.mongodb.MongodbSingleCollectImpl
 com.usthe.collector.collect.snmp.SnmpCollectImpl
 com.usthe.collector.collect.ssh.SshCollectImpl
