+++ manager/src/main/resources/define/param/param-redis.yml
@@ -50,3 +50,9 @@ param:
       en-US: Password
     type: password
     required: false
+  - field: pattern
+    name:
+      zh-CN: 模式
+      en-US: Pattern
+    required: true
+    defaultValue: 1
